Add CLI flag

+++ scripts/generate.sh
@@ -13,18 +13,23 @@ if [[ -n "${FAIL}" ]]; then
     exit $FAIL
 fi
 
-rm ./downloaded/specs/*
+if [[ $1 == "--fetch-specs" ]]; then
+  echo "Removing existing specs in downloaded/specs/\n"
+  rm ./downloaded/specs/*
+fi
 
 for SERVICE in "tenant" "authorizer" "registry" "decision_logs"
 do
-    TARGET=https://api.github.com/repos/aserto-dev/openapi-grpc/contents/publish/${SERVICE}/openapi.json?ref=${COMMIT_HASH}
     OUTPUT_PATH=./downloaded/specs/${SERVICE}.openapi.json
 
-    echo "CURLing $TARGET to $OUTPUT_PATH"
-    curl \
-        --create-dirs -o "${OUTPUT_PATH}" -u "${USERNAME}:${READ_WRITE_TOKEN}" \
-        -H "Accept: application/vnd.github.v3.raw" \
-        "${TARGET}"
+    if [[ $1 == "--fetch-specs" ]]; then
+        TARGET=https://api.github.com/repos/aserto-dev/openapi-grpc/contents/publish/${SERVICE}/openapi.json?ref=${COMMIT_HASH}
+        echo "CURLing $TARGET to $OUTPUT_PATH"
+        curl \
+            --create-dirs -o "${OUTPUT_PATH}" -u "${USERNAME}:${READ_WRITE_TOKEN}" \
+            -H "Accept: application/vnd.github.v3.raw" \
+            "${TARGET}"
+    fi
     
     yarn openapi-typescript ${OUTPUT_PATH} --output ./generated/${SERVICE}.ts
 done
